moved forwarding header modification to `balance` so that modified header is logged as `up` transaction
