+++ forward/handler_test.go
@@ -8,7 +8,6 @@ import (
 	"strings"
 	"testing"
 
-	"crypto/tls"
 	"github.com/ichiban/jesi/cache"
 )
 
@@ -21,215 +20,24 @@ func TestHandler_ServeHTTP(t *testing.T) {
 		expectedReq  *http.Request
 		expectedResp *http.Response
 	}{
-		{ // with hop-by-hop headers
-			givenReq: &http.Request{
-				Method: http.MethodGet,
-				URL: &url.URL{
-					Path: "/foo",
-				},
-				Header: http.Header{
-					"Connection": []string{"foo, bar"},
-					"Foo":        []string{"abc"},
-					"Bar":        []string{"def"},
-					"Baz":        []string{"ghq"},
-				},
-			},
-			givenResp: &http.Response{
-				StatusCode: http.StatusOK,
-				Header: http.Header{
-					"Connection": []string{"hoge, foo"},
-					"Hoge":       []string{"abc"},
-					"Fuga":       []string{"def"},
-					"Foo":        []string{"ghq"},
-				},
-				Body: ioutil.NopCloser(strings.NewReader("foo")),
-			},
-			expectedReq: &http.Request{
-				Method: http.MethodGet,
-				URL: &url.URL{
-					Path: "/foo",
-				},
-				Header: http.Header{
-					"Baz":               []string{"ghq"},
-					"X-Forwarded-Proto": []string{"http"},
-					"Forwarded":         []string{`proto=http`},
-				},
-			},
-			expectedResp: &http.Response{
-				StatusCode: http.StatusOK,
-				Header: http.Header{
-					"Fuga": []string{"def"},
-				},
-				Body: ioutil.NopCloser(strings.NewReader("foo")),
-			},
+		{ // success
+			givenReq:     &http.Request{Method: http.MethodGet, URL: &url.URL{Path: "/foo"}, Header: http.Header{"Foo": []string{"abc"}, "Bar": []string{"def"}}},
+			givenResp:    &http.Response{StatusCode: http.StatusOK, Header: http.Header{"Hoge": []string{"abc"}, "Fuga": []string{"def"}}, Body: ioutil.NopCloser(strings.NewReader("foo"))},
+			expectedReq:  &http.Request{Method: http.MethodGet, URL: &url.URL{Path: "/foo"}, Header: http.Header{"Foo": []string{"abc"}, "Bar": []string{"def"}}},
+			expectedResp: &http.Response{StatusCode: http.StatusOK, Header: http.Header{"Hoge": []string{"abc"}, "Fuga": []string{"def"}}, Body: ioutil.NopCloser(strings.NewReader("foo"))},
 		},
-		{ // with RemoteAddr, Host, and TLS
-			givenReq: &http.Request{
-				RemoteAddr: "192.0.2.1:12345",
-				Host:       "example.com",
-				Method:     http.MethodGet,
-				URL: &url.URL{
-					Path: "/foo",
-				},
-				Header: http.Header{},
-				TLS:    &tls.ConnectionState{},
-			},
-			givenResp: &http.Response{
-				StatusCode: http.StatusOK,
-				Header:     http.Header{},
-				Body:       ioutil.NopCloser(strings.NewReader("foo")),
-			},
-			expectedReq: &http.Request{
-				Method: http.MethodGet,
-				URL: &url.URL{
-					Path: "/foo",
-				},
-				Header: http.Header{
-					"X-Forwarded-For":   []string{"192.0.2.1"},
-					"X-Forwarded-Host":  []string{"example.com"},
-					"X-Forwarded-Proto": []string{"https"},
-					"Forwarded":         []string{`for="192.0.2.1:12345";host=example.com;proto=https`},
-				},
-			},
-			expectedResp: &http.Response{
-				StatusCode: http.StatusOK,
-				Header:     http.Header{},
-				Body:       ioutil.NopCloser(strings.NewReader("foo")),
-			},
-		},
-		{ // with RemoteAddr, X-Forwarded-For, and Forwarded
-			givenReq: &http.Request{
-				RemoteAddr: "192.0.2.3:12345",
-				Method:     http.MethodGet,
-				URL: &url.URL{
-					Path: "/foo",
-				},
-				Header: http.Header{
-					"X-Forwarded-For":   []string{"192.0.2.1, 192.0.2.2"},
-					"X-Forwarded-Proto": []string{"http"},
-					"Forwarded":         []string{`for="192.0.2.1:12345"`, `for="192.0.2.2:12345"`},
-				},
-			},
-			givenResp: &http.Response{
-				StatusCode: http.StatusOK,
-				Header:     http.Header{},
-				Body:       ioutil.NopCloser(strings.NewReader("foo")),
-			},
-			expectedReq: &http.Request{
-				Method: http.MethodGet,
-				URL: &url.URL{
-					Path: "/foo",
-				},
-				Header: http.Header{
-					"X-Forwarded-For":   []string{"192.0.2.1, 192.0.2.2, 192.0.2.3"},
-					"X-Forwarded-Proto": []string{"http"},
-					"Forwarded":         []string{`for="192.0.2.1:12345"`, `for="192.0.2.2:12345"`, `for="192.0.2.3:12345";proto=http`},
-				},
-			},
-			expectedResp: &http.Response{
-				StatusCode: http.StatusOK,
-				Header:     http.Header{},
-				Body:       ioutil.NopCloser(strings.NewReader("foo")),
-			},
-		},
-		{ // with RemoteAddr, X-Forwarded-For (X-Forwarded-For will be converted to Forwarded)
-			givenReq: &http.Request{
-				RemoteAddr: "192.0.2.3:12345",
-				Method:     http.MethodGet,
-				URL: &url.URL{
-					Path: "/foo",
-				},
-				Header: http.Header{
-					"X-Forwarded-For": []string{"192.0.2.1, 192.0.2.2"},
-				},
-			},
-			givenResp: &http.Response{
-				StatusCode: http.StatusOK,
-				Header:     http.Header{},
-				Body:       ioutil.NopCloser(strings.NewReader("foo")),
-			},
-			expectedReq: &http.Request{
-				Method: http.MethodGet,
-				URL: &url.URL{
-					Path: "/foo",
-				},
-				Header: http.Header{
-					"X-Forwarded-For":   []string{"192.0.2.1, 192.0.2.2, 192.0.2.3"},
-					"X-Forwarded-Proto": []string{"http"},
-					"Forwarded":         []string{`for=192.0.2.1`, `for=192.0.2.2`, `for="192.0.2.3:12345";proto=http`},
-				},
-			},
-			expectedResp: &http.Response{
-				StatusCode: http.StatusOK,
-				Header:     http.Header{},
-				Body:       ioutil.NopCloser(strings.NewReader("foo")),
-			},
-		},
-		{ // with RemoteAddr, X-Forwarded-For, and X-Forwarded-By
-			givenReq: &http.Request{
-				RemoteAddr: "192.0.2.3:12345",
-				Method:     http.MethodGet,
-				URL: &url.URL{
-					Path: "/foo",
-				},
-				Header: http.Header{
-					"X-Forwarded-By":  []string{"192.0.2.10"},
-					"X-Forwarded-For": []string{"192.0.2.1, 192.0.2.2"},
-				},
-			},
-			givenResp: &http.Response{
-				StatusCode: http.StatusOK,
-				Header:     http.Header{},
-				Body:       ioutil.NopCloser(strings.NewReader("foo")),
-			},
-			expectedReq: &http.Request{
-				Method: http.MethodGet,
-				URL: &url.URL{
-					Path: "/foo",
-				},
-				Header: http.Header{
-					"X-Forwarded-By":    []string{"192.0.2.10"},
-					"X-Forwarded-For":   []string{"192.0.2.1, 192.0.2.2, 192.0.2.3"},
-					"X-Forwarded-Proto": []string{"http"},
-					"Forwarded":         []string{`for=192.0.2.1`, `for=192.0.2.2`, `for="192.0.2.3:12345";proto=http`},
-				},
-			},
-			expectedResp: &http.Response{
-				StatusCode: http.StatusOK,
-				Header:     http.Header{},
-				Body:       ioutil.NopCloser(strings.NewReader("foo")),
-			},
-		},
-		{ // with error
-			givenReq: &http.Request{
-				Method: http.MethodGet,
-				URL: &url.URL{
-					Path: "/foo",
-				},
-				Header: http.Header{},
-			},
-			givenResp: &http.Response{},
-			givenErr:  errors.New("failed"),
-			expectedReq: &http.Request{
-				Method: http.MethodGet,
-				URL: &url.URL{
-					Path: "/foo",
-				},
-				Header: http.Header{
-					"X-Forwarded-Proto": []string{"http"},
-					"Forwarded":         []string{`proto=http`},
-				},
-			},
-			expectedResp: &http.Response{
-				StatusCode: http.StatusBadGateway,
-				Header:     http.Header{},
-			},
+		{ // error
+			givenReq:     &http.Request{Method: http.MethodGet, URL: &url.URL{Path: "/foo"}},
+			givenErr:     errors.New("failed"),
+			expectedReq:  &http.Request{Method: http.MethodGet, URL: &url.URL{Path: "/foo"}},
+			expectedResp: &http.Response{StatusCode: http.StatusBadGateway, Body: ioutil.NopCloser(strings.NewReader(""))},
 		},
 	}
 
 	for i, tc := range testCases {
 		transport := &testTransport{
 			resp: tc.givenResp,
+			err:  tc.givenErr,
 		}
 		handler := &Handler{
 			Transport: transport,
